refactor: extract catch_all route to dedicated redirect module
- Move ARK redirect logic from main ark.py to routes/redirect.py
- Create redirect_bp blueprint for better code organization
- Maintain all existing functionality and error handling
- Improve code modularity and separation of concerns
- Preparation for parallel execution implementation
